requirements.txt
langchain-community
langchain-google-genai
langchain-huggingface
langchain-core
flask
PyPDF2
python-dotenv